+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This library adheres to [Semantic Versioning](https://semver.org/) and [Keep a CHANGELOG](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Added
+
+* `disable_password_change_notification`: Added a feature that disables sending password change notification emails to site admins.
+
 ## 2.2.1
 
 ### Added