+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This library adheres to [Semantic Versioning](https://semver.org/) and [Keep a CHANGELOG](https://keepachangelog.com/en/1.0.0/).
 
-## Unreleased
+## 2.3.0
 
 ### Added